ObjectDetection/EfficientDet/Tensorflow/source/automl/efficientdet/requirements.txt
absl-py==0.9.0
lxml==4.5.0
matplotlib==3.2.1
numpy==1.18.3
Pillow==7.1.2
PyYAML==5.3.1
six==1.14.0
tensorflow-gpu==2.2.0
